extract & reuse StubInstanceMethodSharedTests
